% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_distrital.R
\docType{data}
\name{mapa_distrital}
\alias{mapa_distrital}
\title{Mapa de límites distritales de Perú}
\format{
A data frame with 1874 rows and 16 variables:
\describe{
\item{departamento}{Department name. First-level administrative division}
\item{provincia}{Province name. Second-level administrative division}
\item{distrito}{District name. Third-level administrative division}
\item{ubigeo}{INEI's geographic location code}
\item{geometry}{Column containing simple features data}
}
}
\source{
\url{https://www.geogpsperu.com/2019/05/limite-distrital-actualizado-inei.html}
}
\usage{
mapa_distrital
}
\description{
A simple features dataset containing the limits of the 1874 peruvian districts. The data set was found in www.geogpsperu.com .
}
\keyword{datasets}
