% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_regional.R
\docType{data}
\name{mapa_regional}
\alias{mapa_regional}
\title{Mapa de límites regionales de Perú}
\format{
A data frame with 25 rows and 3 variables:
\describe{
\item{departamento}{Department name. First-level administrative division}
\item{cod_depa}{Department code. Extracted from INEI's geographic location code}
\item{geometry}{Column containing simple features data}
}
}
\source{
\url{https://www.geogpsperu.com/2019/05/limite-distrital-actualizado-inei.html}
}
\usage{
mapa_regional
}
\description{
A simple features dataset containing the limits of the 25 peruvian provinces. The data set is based on mapa_distrital.
}
\keyword{datasets}
