% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonas_lima.R
\docType{data}
\name{zonas_lima}
\alias{zonas_lima}
\title{Distritos de Lima por zonas}
\format{
A data frame with 43 rows and 3 variables:
\describe{
\item{distrito}{District name. Third-level administrative division}
\item{ubigeo}{INEI's geographic location code}
\item{zona}{District zone}
}
}
\source{
\url{https://www.mapadelima.com/mapa-de-distritos-de-lima/}
}
\usage{
zonas_lima
}
\description{
A  dataset containing the names and zone of the 43 districts of Lima, Perú.
}
\keyword{datasets}
